Revert rename share_object_state_machines.py
